Fix buttons, move TopBar to a designated component
