+++ lib/src/NavigationMock/constants.ts
@@ -1,2 +1,3 @@
 export const VISIBLE_SCREEN_TEST_ID = 'VISIBLE_SCREEN_TEST_ID';
 export const VISIBLE_OVERLAY_TEST_ID = 'VISIBLE_OVERLAY_TEST_ID';
+export const DEFAULT_BACK_BUTTON_TEST_ID = 'RNN.back';
